Adds a Mount decap dungeon
Adds a Mount decap dungeon far off from the start with a lootpile and a ravox ring

puts two strife rings on the map, one in the sewer hidden and another in a wolf pit.

adds a dendor amulet of the hunt for speed buff, not on map

adds a few sprites for the new buffs.